+++ R/data.R
@@ -1,8 +1,9 @@
-#' DisMAP regions, seasons and dataset_code
+#' DisMAP datasets, given by region and optionally season
 #'
 #' A multiline simple feature boundary of regions and seasons used in DisMAP.
 #'
-#' @format ## `dm_regions`
+#' @format ## `dm_datasets`
+#' @concept data
 #' A data frame with 14 features and 4 fields:
 #' - `dataset_code`: DisMAP dataset code
 #' - `region`: Region name
@@ -10,4 +11,4 @@
 #' - `distribution_project_code`: DisMAP distribution project code
 #'
 #' @source <https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/DisMAP_Regions_CURRENT/FeatureServer/1>
-"dm_regions"
+"dm_datasets"