+++ NEWS.md
@@ -1,3 +1,12 @@
+# dismapr 0.4.0
+
+* Added concepts to functions for Reference section: Indicators, Rasters, Surveys, Data.
+* Broke up functions by concept into seperate reference R files.
+* Changed `dm_regions` to `dm_datasets` to enforce hierarchical consistency with DisMAP Rasters:
+  - **datasets**: region and possibly seasons
+  - **layers**: scientific species name or "Species Richness"
+  - **years**: years of data available
+  
 # dismapr 0.3.0
 
 * Made consistent "species" to "layers": `get_dm_dataset_species_years()` -> `get_dm_dataset_layer_years()`.